Update field type and optimize update method to VehicleModel
Changed the data type of 'horse_power' field in VehicleModel to unsignedInteger. In the Vehicle Model Repository, an explicit field list has been provided in the update method to circumvent a possible bug. Further, factory data generation has been enhanced with new fields for better testing scenarios. The field rules in VehicleModelUpdateRequest have also been modified to match these changes.

+++ database/migrations/2024_02_07_102653_add_fields_in_vehicle_models_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::table('vehicle_models', function (Blueprint $table) {
             $table->enum('gearbox', ['automatic','manual'])->after('estimated_price');
             $table->enum('fuel_type', ['gasoline', 'petrol', 'diesel', 'propane', 'compressed natural gas', 'liquefied petroleum gas', 'ethanol', 'methanol', 'biodiesel', 'hydrogen'])->after('gearbox');
-            $table->decimal('horse_power', 10, 2)->after('fuel_type');
+            $table->unsignedInteger('horse_power')->after('fuel_type');
             $table->decimal('consumption', 10, 2)->after('horse_power');
             $table->decimal('release_year', 10, 2)->after('consumption');
         });
